Forward more metadata functions to the inner XLA tensor.
This PR overrides more tensor metadata functions to the inner tensor inside
`FunctionalTensorWrapper`. This is necessary so that we don't have to look at 2 different
metadata, i.e. `FunctionalTensorWrapper`'s and whatever is the `TensorImpl` of its inner
tensor.

We should probably override all `*_custom` functions, but I will leave those for a future
PR. This one takes care of other trivial ones that already have a sym/not-sym counterpart
covered.

ghstack-source-id: a94f755042818710f755a48a2a17e7a8422f7034
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135497