Preserve storage offset on meta conversions.
This PR modifies the `to_meta` function present in the functionalization layer, so that it
also preserves the storage offset.

It does so by creating a base tensor, with enough elements for a view with the source
tensor's metadata to live. Then, we follow that by calling `as_strided`, resulting in a
tensor with the expected metadata.

ghstack-source-id: 8c0c422899d48ec4423fcda2e1c1dc4e652dcd1f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135498